+++ app/examples/src/main/lollypop/MacroDemo.sql
@@ -25,12 +25,6 @@ macro 'jvm' := {
 }
 
 ////////////////////////////////////////////////////////////////////
-// macro ls - List files
-// Usage: ls '-al'
-////////////////////////////////////////////////////////////////////
-macro 'ls %e:options' := os_exec('ls {{ options }}')
-
-////////////////////////////////////////////////////////////////////
 // macro tickers - Generates random stock quotes for testing
 // Usage: val stocks = tickers 100000
 ////////////////////////////////////////////////////////////////////
